requirements.txt
ultralytics>=8.0.0
opencv-python
paho-mqtt
requests
numpy